Prepare for integration of evohome client  - move HGI80 specific code to EvohomeSerial  - move remaining code to EvohomeBase
