+++ CMakeLists.txt
@@ -311,7 +311,8 @@ hardware/EcoDevices.cpp
 hardware/EnOceanESP2.cpp
 hardware/EnOceanESP3.cpp
 hardware/Ec3kMeterTCP.cpp
-hardware/evohome.cpp
+hardware/EvohomeBase.cpp
+hardware/EvohomeSerial.cpp
 hardware/ETH8020.cpp
 hardware/FritzboxTCP.cpp
 hardware/GoodweAPI.cpp
